gfx/internal/util: fix preload failure with PreDraw
`foobarToLoad` was the number of a particular type of resource (like textures)
to load. It was calculated based on whether or not that object was already
loaded, such that `foobarToLoad` was only increased if there were `foobar`
objects _not already loaded_.

After calculating, we would iterate over the objects _but ignore the Loaded
status of the object_ and proceed to ask the device to load it. This meant that
if the second texture of an object was not loaded, but the first one was, then
`texturesToLoad` would be one, and we would ask the device to load the first
texture instead of the second one (which needed to be loaded!). This caused a
data race when rendering because it meant the texture/mesh/etc wasn't loaded.

Helps #150