gfx/gl2: fix data race in resource loading
Without a `gl.Finish` operation we open up the possibility of the rendering
thread to use a texture etc that has not actually been entirely loaded by the
graphics hardware. Fix this by using a `gl.Finish` operation which sync's the
GPU and CPU entirely. This is more aggressive than we need, so #158 has been
opened to address this in the future.